+++ include/unity/scopes/internal/ScopeObjectBase.h
@@ -55,14 +55,14 @@ public:
 
     virtual MWQueryCtrlProxy activate(Result const& result,
                                       ActionMetadata const& hints,
-                                      MWReplyProxy const &reply,
+                                      MWReplyProxy const& reply,
                                       InvokeInfo const& info) = 0;
 
     virtual MWQueryCtrlProxy perform_action(Result const& result,
                                             ActionMetadata const& hints,
                                             std::string const& widget_id,
                                             std::string const& action_id,
-                                            MWReplyProxy const &reply,
+                                            MWReplyProxy const& reply,
                                             InvokeInfo const& info) = 0;
 
     virtual MWQueryCtrlProxy preview(Result const& result,