+++ RELEASE_NOTES.md
@@ -1,9 +1,18 @@
 Release notes
 =============
 
+Changes in version 0.4.6
+========================
+  - Added method to get and set display hints for filters (at this moment the only display hint
+    available is Primary hint).
+
+  - Added has_subdepartments flag and alternate label to Department class.
+
+  - Added TTL option for scope's results.
+
 Changes in version 0.4.5
 ========================
-  * Implemented RangeInputFilter.
+  - Implemented RangeInputFilter.
 
 Changes in version 0.4.4
 ========================